Update Replicated and Docker Versions
Replicated 2.29.0 and Docker 17.12.1-ce

Our primary reason for upgrading Replicated is that newer versions
should be able to pull all of the Docker images in parallel which will
greatly speed up the installation time. It's also good to stay
up-to-date with bugfixes and features though.

The version of Docker should be changed at the same time to match
Replicated's recommendation:

https://help.replicated.com/docs/native/customer-installations/installing-manually/