Upgrade Replicated to 2.29.0 and Docker to 17.12.1
We are currently using Replicated version 2.10.3 and Docker version 17.03.2 which are both very old.  We are upgrading Replicated to 2.29.0 to take advantage of  many bug fixes.  They have also switched from using piper for Docker pulls to using the native Docker pull client which allows for parallel pulls.  This hopefully will allow for faster tests.

Replicated recommends using Docker version 17.12.1 with the latest version of Replicated so we are updating it as well.